[Metricbeat][iostat] Add queue time metric
This commit adds the queue time metric as computed by pt-diskstats.

Queue time is the time a request spends in the OS queue before it is
sent to the physical device. The metric gives an average of the queue
time over the collection interval.

Other existing metrics like await or aqu-sz (average queue size) don't
provide visibility into the OS queue only. They conflate the time spent
in the OS queue with the time spent in the physical device queue.